fix(pulseaudio,pipewire): build on targets without 64-bit atomics
PipeWire's per-stream quantum tracking never needs more than 32 bits
(a callback's frame count is bounded by its mapped SPA buffer size,
itself a uint32_t), so narrow last_quantum to AtomicU32 and drop the
dependency entirely there.

PulseAudio's latency tracking genuinely needs 64 bits, so pull in
portable-atomic on targets without native support for it.

Add a CI job on mipsel to catch regressions on 32-bit-atomic targets.

+++ src/host/pipewire/stream.rs
@@ -3,7 +3,7 @@ use std::{
     rc::Rc,
     sync::{
         Arc, Mutex,
-        atomic::{AtomicBool, AtomicU64, Ordering},
+        atomic::{AtomicBool, AtomicU32, Ordering},
     },
     thread::JoinHandle,
     time::Instant,
@@ -83,7 +83,7 @@ pub(super) enum StreamCommand {
 pub struct Stream {
     handle: Option<JoinHandle<()>>,
     controller: pw::channel::Sender<StreamCommand>,
-    last_quantum: Arc<AtomicU64>,
+    last_quantum: Arc<AtomicU32>,
     start: Instant,
     latch: Latch,
     is_output: bool,
@@ -96,7 +96,7 @@ impl Stream {
     pub(super) fn new(
         handle: JoinHandle<()>,
         controller: pw::channel::Sender<StreamCommand>,
-        last_quantum: Arc<AtomicU64>,
+        last_quantum: Arc<AtomicU32>,
         start: Instant,
         latch: Latch,
         is_output: bool,
@@ -287,7 +287,7 @@ pub struct UserData<D> {
     error_callback: ErrorCallbackArc,
     sample_format: SampleFormat,
     format: AudioInfoRaw,
-    last_quantum: Arc<AtomicU64>,
+    last_quantum: Arc<AtomicU32>,
     start: Instant,
     draining: Arc<AtomicBool>,
     is_default_device: bool,
@@ -380,14 +380,14 @@ where
     ) {
         #[cfg(feature = "realtime")]
         {
-            let prev = self.last_quantum.swap(frames as u64, Ordering::Relaxed);
-            if !self.rt_promoted || frames as u64 != prev {
+            let prev = self.last_quantum.swap(frames as u32, Ordering::Relaxed);
+            if !self.rt_promoted || frames as u32 != prev {
                 self.promote_realtime(frames as FrameCount);
             }
         }
 
         #[cfg(not(feature = "realtime"))]
-        self.last_quantum.store(frames as u64, Ordering::Relaxed);
+        self.last_quantum.store(frames as u32, Ordering::Relaxed);
 
         if !self.draining.load(Ordering::Relaxed) {
             let (callback, capture) = match pw_stream_time(stream) {
@@ -432,14 +432,14 @@ where
     ) {
         #[cfg(feature = "realtime")]
         {
-            let prev = self.last_quantum.swap(frames as u64, Ordering::Relaxed);
-            if !self.rt_promoted || frames as u64 != prev {
+            let prev = self.last_quantum.swap(frames as u32, Ordering::Relaxed);
+            if !self.rt_promoted || frames as u32 != prev {
                 self.promote_realtime(frames as FrameCount);
             }
         }
 
         #[cfg(not(feature = "realtime"))]
-        self.last_quantum.store(frames as u64, Ordering::Relaxed);
+        self.last_quantum.store(frames as u32, Ordering::Relaxed);
 
         if !self.draining.load(Ordering::Relaxed) {
             let (callback, playback) = match pw_stream_time(stream) {
@@ -624,7 +624,7 @@ pub struct ConnectParams {
     pub config: StreamConfig,
     pub properties: PropertiesBox,
     pub sample_format: SampleFormat,
-    pub last_quantum: Arc<AtomicU64>,
+    pub last_quantum: Arc<AtomicU32>,
     pub start: Instant,
     pub connect_automatically: bool,
     pub draining: Arc<AtomicBool>,
